gmp: Force use of alloca for small temp storage
Or at least I think that's what this will do. By default, gmp will try
to use alloca and will fall back to reentrant malloc. This way, if
alloca can't be use for some reason there should be some sort of
notification, in theory. Not like I expect that to happen, though.